refactor: reorganize folders in <features> and <core> and create basic home view
