+++ lib/bootstrap.dart
@@ -4,8 +4,8 @@ import 'dart:io';
 
 import 'package:bloc/bloc.dart';
 import 'package:flutter/widgets.dart';
-import 'package:sq1_assignment/network_service/network_service.dart';
-import 'package:sq1_assignment/service_locator/service_locator.dart';
+import 'package:sq1_assignment/core/di/service_locator.dart';
+import 'package:sq1_assignment/core/network/network_service.dart';
 
 class AppBlocObserver extends BlocObserver {
   const AppBlocObserver();
@@ -29,8 +29,6 @@ Future<void> bootstrap(FutureOr<Widget> Function() builder) async {
   };
 
   Bloc.observer = const AppBlocObserver();
-
-  // Add cross-flavor configuration here
 
   WidgetsFlutterBinding.ensureInitialized();
