+++ lib/core/network/google_maps_api/google_maps_api_client.dart
@@ -2,7 +2,7 @@ import 'package:dio/dio.dart';
 import 'package:dio_cache_interceptor/dio_cache_interceptor.dart';
 import 'package:dio_cache_interceptor_hive_store/dio_cache_interceptor_hive_store.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:sq1_assignment/network_service/network_service.dart';
+import 'package:sq1_assignment/core/network/network_service.dart';
 
 class GoogleMapsApiClient extends NetworkServiceClient {
   GoogleMapsApiClient._internal(List<Interceptor> interceptors)
