+++ lib/core/network/google_maps_api/resource/resource.dart
