+++ lib/core/network/pagination/model/model.dart
