+++ lib/core/network/square1_api/resource/city_resource.dart
