+++ lib/core/network/square1_api/resource/resource.dart
