+++ lib/feature/city_search/cubit/city_search_cubit/city_search_cubit.dart
@@ -1,11 +1,11 @@
 import 'package:bloc/bloc.dart';
 import 'package:collection/collection.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:sq1_assignment/core/network/network_service.dart';
 import 'package:sq1_assignment/feature/city_search/city_search.dart';
-import 'package:sq1_assignment/network_service/network_service.dart';
 
-part 'city_search_state.dart';
 part 'city_search_cubit.freezed.dart';
+part 'city_search_state.dart';
 
 class CitySearchCubit extends Cubit<CitySearchState> {
   CitySearchCubit({
