+++ lib/feature/city_search/data/repository/api_city_search_repository.dart
@@ -1,6 +1,6 @@
 import 'package:multiple_result/multiple_result.dart';
+import 'package:sq1_assignment/core/network/network_service.dart';
 import 'package:sq1_assignment/feature/city_search/city_search.dart';
-import 'package:sq1_assignment/network_service/network_service.dart';
 
 class ApiCitySearchRepository implements CitySearchRepository {
   ApiCitySearchRepository({
