+++ lib/feature/city_search/domain/repository/city_search_repository.dart
@@ -1,6 +1,6 @@
 import 'package:multiple_result/multiple_result.dart';
+import 'package:sq1_assignment/core/network/network_service.dart';
 import 'package:sq1_assignment/feature/city_search/city_search.dart';
-import 'package:sq1_assignment/network_service/network_service.dart';
 
 abstract class CitySearchRepository {
   Future<Result<PaginatedData<City>, String>> getCitiesByName({
