+++ lib/feature/city_search/view/city_search_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:sq1_assignment/core/di/service_locator.dart';
 import 'package:sq1_assignment/feature/city_search/city_search.dart';
-import 'package:sq1_assignment/service_locator/service_locator.dart';
 
 class CitySearchPage extends StatelessWidget {
   const CitySearchPage({super.key});
