--- lib/feature/counter/counter.dart
@@ -1,2 +1,0 @@
-export 'cubit/counter_cubit.dart';
-export 'view/counter_page.dart';
