--- lib/feature/counter/cubit/counter_cubit.dart
@@ -1,8 +1,0 @@
-import 'package:bloc/bloc.dart';
-
-class CounterCubit extends Cubit<int> {
-  CounterCubit() : super(0);
-
-  void increment() => emit(state + 1);
-  void decrement() => emit(state - 1);
-}
