+++ lib/feature/home/view/screens/home_screen.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class HomeScreen extends StatelessWidget {
+  const HomeScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+      body: const Center(
+        child: Text('Home'),
+      ),
+    );
+  }
+}
