+++ test/feature/city_search/cubit/city_search_cubit/city_search_cubit_test.dart
@@ -3,8 +3,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:multiple_result/multiple_result.dart';
+import 'package:sq1_assignment/core/network/network_service.dart';
 import 'package:sq1_assignment/feature/city_search/city_search.dart';
-import 'package:sq1_assignment/network_service/network_service.dart';
 
 import 'city_search_cubit_test.mocks.dart';
