+++ test/feature/city_search/cubit/city_search_cubit/city_search_cubit_test.mocks.dart
@@ -9,7 +9,7 @@ import 'package:mockito/mockito.dart' as _i1;
 import 'package:mockito/src/dummies.dart' as _i6;
 import 'package:multiple_result/multiple_result.dart' as _i4;
 import 'package:sq1_assignment/feature/city_search/city_search.dart' as _i2;
-import 'package:sq1_assignment/network_service/network_service.dart' as _i5;
+import 'package:sq1_assignment/core/network/network_service.dart' as _i5;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
@@ -40,33 +40,30 @@ class MockCitySearchRepository extends _i1.Mock
     int? page,
   }) =>
       (super.noSuchMethod(
-            Invocation.method(#getCitiesByName, [], {#name: name, #page: page}),
-            returnValue: _i3.Future<
-              _i4.Result<_i5.PaginatedData<_i2.City>, String>
-            >.value(
-              _i6.dummyValue<_i4.Result<_i5.PaginatedData<_i2.City>, String>>(
-                this,
-                Invocation.method(#getCitiesByName, [], {
-                  #name: name,
-                  #page: page,
-                }),
-              ),
-            ),
-          )
-          as _i3.Future<_i4.Result<_i5.PaginatedData<_i2.City>, String>>);
+        Invocation.method(#getCitiesByName, [], {#name: name, #page: page}),
+        returnValue:
+            _i3.Future<_i4.Result<_i5.PaginatedData<_i2.City>, String>>.value(
+          _i6.dummyValue<_i4.Result<_i5.PaginatedData<_i2.City>, String>>(
+            this,
+            Invocation.method(#getCitiesByName, [], {
+              #name: name,
+              #page: page,
+            }),
+          ),
+        ),
+      ) as _i3.Future<_i4.Result<_i5.PaginatedData<_i2.City>, String>>);
 
   @override
   _i3.Future<_i4.Result<_i2.Location, String>> getCityLocation(
     _i2.City? city,
   ) =>
       (super.noSuchMethod(
+        Invocation.method(#getCityLocation, [city]),
+        returnValue: _i3.Future<_i4.Result<_i2.Location, String>>.value(
+          _i6.dummyValue<_i4.Result<_i2.Location, String>>(
+            this,
             Invocation.method(#getCityLocation, [city]),
-            returnValue: _i3.Future<_i4.Result<_i2.Location, String>>.value(
-              _i6.dummyValue<_i4.Result<_i2.Location, String>>(
-                this,
-                Invocation.method(#getCityLocation, [city]),
-              ),
-            ),
-          )
-          as _i3.Future<_i4.Result<_i2.Location, String>>);
+          ),
+        ),
+      ) as _i3.Future<_i4.Result<_i2.Location, String>>);
 }
